Make shade plugin compatible with Spring and signed libs
Shade plugin unpacks all the dependencies, and packs them all into a
single big jar file. What if 2 dependencies had the same file (like
Spring libs). Or what if one of the dependencies was signed? Configured
shade-plugin to be compatible with Spring and signed libs.

In the future we should look into other ways of handling this (packing
jar files inside the uber jar file) - it'd work faster and won't have
the aforementioned problems.